runenr e tabelle funzionanti

+++ EnergIApp/src/main/java/it/epicode/energiapp/services/UserService.java
@@ -54,7 +54,7 @@ public class UserService {
     }
 
     @Transactional
-    public User updateUser(int id, UserRegisterRequestPayloadDTO userDto){
+    public User updateUser(long id, UserRegisterRequestPayloadDTO userDto){
         Optional<User> userOptional = getUserById(id);
 
         if (userOptional.isPresent()){
